requirements.txt
# ── Core API ─────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12

# ── LangChain (Groq compatible) ──────────────────────────────
langchain==0.3.3
langchain-community==0.3.2
langchain-groq==0.2.1

# ── Vector Database ──────────────────────────────────────────
faiss-cpu==1.8.0

# ── Embeddings (Free) ────────────────────────────────────────
sentence-transformers==3.1.1

# ── NLP / Sentiment ──────────────────────────────────────────
textblob==0.18.0.post0

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.2